requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
alembic>=1.10.0
pymysql>=1.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
boto3>=1.26.0
requests>=2.28.0
pandas>=1.5.0
openpyxl>=3.0.0
celery>=5.2.0
redis>=4.0.0
python-dateutil>=2.8.0
cryptography>=39.0.0
loguru>=0.6.0 